refactor: decompose code, add common refactorings
